lockfile: add SRPM name in metadata field
As part of Bodhi gating, we need to deal with source RPM names and not
binary RPMs. E.g. Bodhi CI messages include Koji NVRs, not binary RPMs.
Having to do the translation between SRPM and RPM is expensive because
you need either repodata or e.g. bring up a VM and query the rpmdb.

Instead, just have rpm-ostree output the SRPM name as a metadata field
in the output lockfile. cosa already saves this and outputs it in the
build dir so it ends up in S3 and is much easier to work with.

Obviously there's some redundancy there since every binary RPM from
the same source RPM will have the same field. But it doesn't seem worth
extending the schema for it instead of just working with it as is.

Related: https://github.com/coreos/fedora-coreos-tracker/issues/1617