tmpfiles: Change `read_tmpfiles` to return a direct hashmap
While returning an iterator was more elegant, it makes
error handling much more difficult *and* we had intermediate
`collects()` that meant it wasn't really lazy.

Trying to remove the intermediate `collect` runs into borrowing
issues (on the input `&Dir`) and this isn't a high performance
path so doing the more obvious simpler thing is better.