core: Don't touch /usr/local at assembly time
In a server-side compose, we call `rootfs_prepare_links()` twice:
once as part of unified core assembly, and once as part of final
postprocessing. In case of the former, all we really want are the `/var`
compat symlinks before running scriptlets.

Otherwise, let's reduce to a single call the place where we determine
the fate of `/usr/local` server-side. And certainly client-side, we
shouldn't touch it at all (unless it's part of e.g. some experimental
knob that purposely does more invasive things).

A follow up to this is to split out `/usr/local` handling entirely
into a separate function call, and only call that function in the
server-side compose path (and rename `rootfs_prepare_links()` to e.g.
`rootfs_prepare_compat_var_symlinks()`).