core: Add knob to make it easier to test out state overlay model
I want people to be able to test out the state overlay approach without
having to switch over the base compose in e.g. FCOS to use it.

Hackily add an env var that will forcibly fix the symlinks in the OSTree
commit when doing a local assembly. This is enough to allow properly
checking out RPMs that install into `/opt` and `/usr/local`. Have it also
enable the systemd services for the overlays. (In `/usr`, so that if the
knob is removed, the services are no longer enabled.)