importer: Use rpm functionality to dup fd
I'm hitting a failure in this code when using podman machine on MacOS,
and using coreos-assembler with the `/srv` being a virtiofs mount
back to the host.  Ultimately it looks like `read()` on the fd
for the package is returning `EIO`.

I suspect what's happening here is virtiofs (whether host or guest)
isn't quite handling this scenario where the file is unlinked
on the host.

This patch doesn't help...but it does clean things up here
a bit anyways because librpm has an API for this, we don't
need to indirect through /proc/self/fd.