libpriv: Stop digging in private libcap internals
When handling file caps, we need to be able to get the actual xattr
bytes. Currently, libcap doesn't provide a public API for this.
Meanwhile, we've been hacking around this by redefining internal libcap
structs and using internal fields. This is extremely brittle since any
change to the struct will break us.

And this is exactly what happened in the libcap currently in rawhide. A
new field was added to the struct:

https://git.kernel.org/pub/scm/libs/libcap/libcap.git//commit/?id=aca076443591ba18438b60e41294b59a324daf04

This patch works around the lack of a public API a different way: we
allocate an `O_TMPFILE`, set caps on it, then read back the xattr. This
is more heavyweight but few binaries actually have file caps anyway. It
also requires `CAP_SETFCAP`, but that's already the case today since the
assembly code also applies them to checked out files on-disk.

No new tests for this; we have existing coverage for file caps. That
coverage would've caught this if we ran CI against rawhide.

Closes #4765