docs/treefile.md: Document postprocess script ordering
The semantics around this are a bit confusing and I had to rederive
them again today. Ideally we'd change it to something less surprising
(specifically the second rule) but that'd be a breaking change. Let's at
least try to document it well.

Obviously these rules apply to any list we merge, but postprocess
scripts in particular are imperative bits of code and may affect each
other and so it's not surprising that one would care more about the
semantics for those vs e.g. `packages`. So let's document it there.