container: Add spinner/progress for layer fetches
This at least gives us a spinner when we're fetching
layers, so folks know something is happening.

As the comment says we *can* get byte-level progress, but
it just needs more plumbing work with our complex internal
output infrastructure.

Note that I had to move the layer progress bits into a
tokio LocalSet because the progress output stuff is
a thread local.

Signed-off-by: Colin Walters <walters@verbum.org>