packaging: drop `.in` extension on `rpm-ostree.spec.in`
There's no processing done on this spec file nor is it hooked to the
buildsystem. Drop the `.in` extension.

Motivated by the fact that `dnf builddep rpm-ostree.spec.in` in dnf5
gets confused by the non `.spec` extension. And it doesn't have a
`--spec` switch yet to disambiguate.

Related: https://github.com/rpm-software-management/dnf5/pull/810