packaging: Use git timestamp as mtime for vendored files
When generating distribution tarballs the creation time of the
tarball ends up in the tarball itself as the stored mtime of
the files created when vendoring any Rust dependencies.

Use the `--mtime` argument to tar to replace the mtime of the
vendored files (and, by extension, Cargo.lock) with the timestamp
of the commit being used to generate the tarball.