Extend rpm-ostreed.service idle-timeout
When making D-Bus calls to the rpm-ostreed.service, the service will
always exit after the specified idle-timeout has passed from when the
daemon was first created. Even if D-Bus calls are constantly being made,
the idle-timeout will still be enforced.

This commit subscribes to the bus, listening for new calls. If a call is
detected, the idle-timeout is reset. The service will only stop after
the idle-timeout has passed after the last call to the bus.

Signed-off-by: Luke Yang <luyang@redhat.com>