rust/bwrap: log `fusermount -u` errors to stderr
I believe cosa is hitting an error in this code path, but we can't
inspect the error because there's no journald running in supermin:

https://github.com/coreos/coreos-assembler/issues/3848

We don't gain much by printing this error to the journal since it
shouldn't be very common. Print it on stderr instead where it's more
likely to have visibility.