sysroot: Add total layer count to output
It's useful to see in the progress output how many layers
there are to fetch.

This is similar to
https://github.com/containers/bootc/pull/524/commits/6eb57180a3c82393b70edb702fb9c4db53e54c32
which ended up being totally reworked in a nicer way in
https://github.com/containers/bootc/pull/655/commits/d8b5df2f378b54c3a60ccaa673a60a7bbca5d653

But doing the latter would require nontrivial changes to our
DBus API around status and progress reporting...and I'd
like to think about how we tackle that more generally
in e.g. https://github.com/containers/skopeo/issues/658

Closes: https://github.com/coreos/rpm-ostree/issues/5024