compose: Initial kickstart support
The treefile syntax is...ok, but it is bespoke to us and
we never really emphasized its use outside of our sphere.

As I'm looking at https://gitlab.com/fedora/bootc/tracker/-/issues/32
I'd really like to expose something that has existing widespread
use as a default entrypoint.

kickstart is really old, long predating even virtualization.

It has a decent set of features (includes, comments) as a baseline
specifically as a mechanism to define package sets, it has key
features we want:

- turning off weak dependencies
- excludes

This is just a stub kickstart parser...the next step is to
wire this up into the treefile.

Signed-off-by: Colin Walters <walters@verbum.org>