daemon/transaction-types: Support custom origins for digested pullspecs
Having a digested pullspec is the container flow equivalent of an
ostree checksum origin. So having a custom origin in that case is useful
there for the same reasons (i.e. it means that something else is likely
managing the host and we want to reflect that info in the origin).

We're thinking of using this in FCOS when we move from OSTree to OCI
for updates.