compose: Allow missing `repos` key
In fedora-bootc, we want the API to set the Fedora version to use to
be `--from quay.io/fedora/fedora:<N>`, from which the repo files are
copied. So in that case, we just want the default repos that are enabled
in that base image to be used. Support this use case in rpm-ostree by
just defaulting to that when `repos` is not specified.

Related: https://gitlab.com/fedora/bootc/tracker/-/issues/39