Add `rpm-ostree experimental`
We already have `rpm-ostree ex` which is the same thing. However,
adding things to that requires touching C code, which we often
immediately turn back around and have invoke Rust, which is
super annoying to juggle.

This CLI code is Rust from the start which makes adding commands
more ergonomic.

For now this is just a placeholder.

Signed-off-by: Colin Walters <walters@verbum.org>