core: Create usr/sbin -> bin if we detect merged sbin filesystem
The filesystem package in Fedora is *absoulutely* insane; it
basically reimplements itself in lua because it can't
rely on being first in the transaction, when it'd be infinitely
saner to just define it to be first (via a metadata flag, whatever).

Anyways...because we don't run that lua code, and for probably
even crazier upgrade reasons the sbin link is only made in lua,
detect when the package has the provides, and create the link.

Now that I think about it we probably never needed the provides,
and we could have just unconditionally made the link if it wasn't
already a directory.

But anyways, using the provides seems cleaner and scopes the
blast radius down a bit.

xref: https://gitlab.com/fedora/bootc/tracker/-/issues/29