libpriv/rpm: Always copy rpmdb even in user mode checkouts
I'm working on doing container reencapsulation and using
a user-mode repo and got burned by the hardlinking here
and rpm always *writing* to the repo even when reading it.

Most systems we care about have reflinks so it's cheap. And in practice
most usage of this before was actually checking out from bare
(not bare-user) repos, so we were copying anyways.

Signed-off-by: Colin Walters <walters@verbum.org>