Use ostree-ext from bootc
The idea is ostree-ext is deprecated in favor of bootc; however
we still directly use it here, and fixing that is going to take
a while longer than I thought.

Switch to pulling a pinned version from bootc's git, so
at least we don't need to maintain a formal crate for it
and think about API stability.

Signed-off-by: Colin Walters <walters@verbum.org>