core/cliwrap: Fix is_ostree_layout() in unit tests
The `is_ostree_layout()` function parsed the real root filesystem.
But we started using it in *unit tests* which creates unpredictable
behavior.

The unit test for the core script wrapping was failing for me,
and digging in I don't see how it could have ever passed. I don't
think we're actually running the Rust tests in CI...need to dig
into that.

- Change `is_ostree_layout` to take a cap-std root
- Add a unit test for it alone
- Change all users to pass the rootfs
- Fix the `undo()` method to undo the kernel-install wrapping
  correctly

With this the unit test passes again, and we shouldn't
have any behavior which depends on the development root.

Signed-off-by: Colin Walters <walters@verbum.org>