tests/compose: Canonicalize perms for root
For some reason I don't understand at all, there's a race
condition in this code where *sometimes* the permissions of
the temporary root are 0777 when they should be 0755. Digging
in only slightly I think this may actually be a refcounting bug
in the `CANONICAL_PERMISSIONS` code in ostree?

But anyways, we expect most people to use sensible mode bits
so let's start out with mode 0755 in a `root` directory in
the tempdir.

Signed-off-by: Colin Walters <walters@verbum.org>