kernel_install: unwrap systemd if it's wrapped
when replacing the kernel packages using rpm-ostree but
with the layout in /usr/lib/kernel/install.conf set to ostree
we need to check for a wrapped systemd before calling dracut
to get the correct modules in the initramfs.

When rpm-ostree kernel-install is called from dnf or rpm we
don't wrap before running scripts so we were ok in that path.