Add new `treefile-apply` experimental command
We are working towards rebasing FCOS and RHCOS on top of fedora-bootc
and rhel-bootc. As one can imagine, we are quite heavily invested in
treefiles currently, so being able to keep using what we already have as
we transition into becoming a layered build is valuable.

Dockerfiles are great of course as the common language used to build
container images. But it's not great for implementing complex logic.

Add a new `rpm-ostree experimental compose treefile-apply` command
which takes a treefile, but applies its directives _live_ in the running
environment.

So in this model, the business logic in the Dockerfile is kept to a
minimum and all the heavy lifting happens via
`RUN rpm-ostree compose treefile-apply manifest.yaml`.

A very small subset of the treefile spec is actually supported. The goal
is _not_ to reimplement everything, but only what is needed and useful
in a derivation context.

For now, only `packages`, `recommends`, and `postprocess` are supported.
More will be added in the future.

This code actually started as a Python script in openshift/os, currently
used to build the OCP node image:

https://github.com/openshift/os/blob/master/scripts/apply-manifest

So in fact, the node image (which in turn is a layered build on top of
RHCOS) will also be using this code.