core: Ignore xattrs in copyup for rpmfi overrides
Ignore xattrs when doing a copyup from the repo.
As far as I can tell and think of, this should always be
safe. Basically we care about xattrs like
`security.selinux` and `security.capabilty`. For SELinux,
we always relabel at the end when generating the commit.
For `security.capability`, it's defined by the RPM.
Other xattrs we shouldn't default to inheriting
from the "base" object - the RPM should set them
in general.

This enables using `rpm-ostree compose` in a podman/docker
container which applies e.g. `container_t` to all files.

Signed-off-by: Colin Walters <walters@verbum.org>