core: Use source root for repos and dnf vars
The goal here is to have the new (more container-native) flow
for `experimental compose rootfs` to always operate on a "repos container"
that we've mounted.

I noticed that `rpmostree_find_and_download_packages()` has
similar behavior, but that's apparently only used in override
flows?

The only ugly workaround we need now is for the rpm-gpg stuff.

Signed-off-by: Colin Walters <walters@verbum.org>