ci/test-container.sh: Run `treefile-apply` test earlier
Otherwise, it runs after the `cliwrap` test, which wraps dnf, but we
want the real dnf here and it's awkward to try to work around that.

The `cliwrap` stuff should probably be tested in a separate run until
we remove suppor for it entirely so it doesn't pollute the rest of the
test environment.