compose: Simplify compose-image entrypoint
The argv juggling is error prone and confusing. We were adding
an argument on the C++ side that we then skipped on the rust side.
Change things so we directly pass the C argv to Rust, re-adding
the subcommand to the arguments on the Rust side so the Clap parser
works.