compose-rootfs: Don't output ostree-container format
Just output a "plain" fsroot. We can't do ostree-container
because (unlike compose image) we won't have the ostree.final-diffid indicator in the output image,
so ostree/bootc won't honor all the ostree stuff.

Combined with a prior bug where we were missing setting
the `security.capability` flag, this was causing the failures
in trying to do the centos-bootc build.

Closes: https://github.com/coreos/rpm-ostree/issues/5314

Signed-off-by: Colin Walters <walters@verbum.org>