compose-rootfs: Fix perms for /
The trick we did here in working around `compose install`
generating a `rootfs` subdirectory by moving its contents into
the parent meant we lost the mode bits.

Somehow, only when targeting this for CentOS we hit on the
mode being 0700 (I think this may relate to umask? We probably
need to audit our whole codebase to avoid umask issues; but
really in the longer term switch to trusting the `filesystem`
rpm instead of manually extracting the root).

Signed-off-by: Colin Walters <walters@verbum.org>