daemon,app: Support rebasing/deploying specific digests in OCI path
In FCOS, we want to keep a tagged refspec in the origin, but still be
able to deploy a specific image.

Enable this.

(It's not quite the same as the OSTree case of course, because there's
no concept of branches there. But at least we can enforce that the
digest is located in the same repo as the canonical refspec.)

This is tracked on the bootc side in
https://github.com/containers/bootc/issues/1165.