daemon: allow rebasing to the same refspec if there is a digest override
https://github.com/coreos/rpm-ostree/pull/5325
introduced the ability to rebase to a specific
digest while keeping the tagged refspec in the
origin.

In FCOS, we update from a pinned image to another
on the same stream (i.e. the same tagged refspec).

Allow this by checking if the digest override is
set.

See https://github.com/coreos/rpm-ostree/pull/5325#issuecomment-2706048328