README.md: Revert deprecation notice, update arch diagram
The deprecation warning seems to be causing more confusion than help right now.
We're going to keep maintaining this project for a long time because
people are using it and it's shipped in products.

The main original goal of the notice was to ensure people didn't
show up out of the blue and start proposing big new changes, but
in reality that's unlikely to happen anyways. The "promoting
where things are happening" is already done organically by just,
well, seeing where things are happening.

While we're here, switch the manual ASCII art to mermaid.js and
add bootc at the toplevel to it, peering with ostree.