treefile-apply: Add `--var` option
When using `treefile-apply`, we want to be able to pass the same
treefile when building for different versions. E.g. for FCOS, Fedora
majors, and for openshift/os RHEL vs CentOS Stream (and 9 vs 10).

So we'll be using conditional includes a lot as we currently do. But
then what would be nice is to be able to set some of the variables at
invocation time so that we don't have to make all these little wrapper
treefiles that just set vars.

I think a common pattern then will be e.g.

```
. /etc/os-release
rpm-ostree experimental compose treefile-apply manifest.yaml --var ID=$ID
```

Tempting to directly add a e.g. `--var-from-os-release` to make that
even easier, but let's stop here for now.