treefile-apply: Handle `repos` key
We have this incredibly complex script in openshift/os whose sole
purpose is downloading the repo files from OpenShift CI:

https://github.com/openshift/os/blob/9bf1802a3b5d/ci/get-ocp-repo.sh

The source of that complexity is that it's called for so many different
purposes, with the expectation that it downloads and enables only the
repos needed for each case.

But this essentially duplicates the repo enablement knobs that already
exist in the treefile (for base composes) and extensions manifest. The
main gap then is the treefile used for layered node image builds.

Add support for the `repos` key for this. If it's not set, use the
default repo enablement. If it's set, override it.

Note though that the model is still very much to use the repos from the
default `/etc/yum.repos.d` location.