core: Process ostree layers before running sysusers
At least in FCOS, we add sysusers dropins via overlays and we need those
already present when running `systemd-sysusers`.

Looking at the history here, there's no reason it seems why we were
only processing ostree layers in the path where we had RPM packages
to overlay. It should be safe to move out of that gate. Anyway on the
client side nothing is hooked to that so it still remains a server-side
feature.

Closes: #5353