core: Ignore replaced files in rpmdb write transaction
We hit an interesting obscure case in cosa recently. After cosa was
upgraded to f42, composing c9s would fail with:

```
error: Installing packages: Writing rpmdb: Error running transaction:
file /usr/sbin/pidof conflicts between attempted installs of
procps-ng-3.3.17-14.el9.x86_64 and procps-ng-3.3.17-14.el9.x86_64
```

Notice the two NEVRAs are the same. How can a package conflict with
itself? procps-ng ships both `/usr/bin/pidof` and `/usr/sbin/pidof`, the
latter being a symlink to the former. When writing the rpmdb, we set the
librpm root dir to `/` (because we don't want to invoke a `chroot()`).

Since in f42 the bin-sbin merge happened, the `/usr/sbin` in the
cosa container is a symlink to `bin`. This means that when librpm
handles `/usr/sbin/pidof`, it thinks that it must be the same file as
`/usr/bin/pidof`. Yet, the payload declares one as a symlink and the
other as a regular file.

Thus, a conflict.

The base error here really is that we call `rpmtsSetRootDir(ts, "/")` at
all instead of passing the target rootfs. We could change this, but it
would be quite disruptive at this point to require `CAP_SYS_CHROOT`. (At
least cosa and base bootc composes wouldn't be affected I think, but I'm
really not confident about all other users.)

An alternative to this is to bwrap into the tmprootfs and somehow invoke
the target rootfs' rpm to write the rpmdb. This would also simplify the
"target rpmdb format" work that was needed with the sqlite switch. But
it would be non-trivial to massage through all the state we need.

A much simpler hack is to just tell librpm to ignore replaced files.
We don't really need this check here. libostree implicitly already did
this when we checked out the packages. The main thing we need from this
transaction is the writing of the rpmdb.

Worth noting also that AFAIK at this point, CoreOS is the odd one
out consistently doing "cross-composes". E.g. if we were using a c9s
source, we wouldn't have had this issue either. Bootc base images are
not cross-composed so wouldn't hit this. And obviously for CoreOS at
least, once we move to building from bootc base images, we'd no longer
be running this code. So this makes it even less interesting to try to
do something more correct but more involved here.

Closes: https://github.com/coreos/coreos-assembler/issues/4083