rust/bwrap: don't swallow STDERR when running commands
Let's merge it with STDOUT so we still get the STDERR logs when
running commands.

The reason I came upon this is because when we run rpm-ostree compose
postprocess scripts we were clearly not getting all of the output. For
example a lot of our scripts are `set -x` but none of that output was
getting printed anywhere, and also things like `systemctl preset-all`
wouldn't show any output at all (which apparently that commands writes
its output to STDERR for some reason).