compose: Support separate cache dir for dnf content
Split `cachedir_dfd` into two: `dnf_cachedir_dfd` and
`ostree_cachedir_dfd`.

Currently, those are effectively the same dfd, retaining current
behaviour. But we add a new (for now unused) switch that allows them
to differ. When active, `ostree_cachedir_dfd` is used for the pkgcache,
the build repo and workdir, and `dnf_cachedir_dfd` is used for caching
the libdnf parts, but notably also opting into the `keepcache` option to
retain RPMs.