compose: Cache RPMs if cache dir on separate filesystem
When using `rpm-ostree compose rootfs` as done by the bootc custom base
image flow within a container build, the output rootfs directory needs
to be on the build container's overlay rootfs so it's usable to the
next stage.

This then conflicts with trying to use a separate mounted cache
directory since the rootfs contains hardlinks from the pkgcache repo.

Let's handle this case gracefully by detecting it and storing only
libdnf bits in the cache directory, but crucially also preserving RPMs.
The pkgcache then is created elsewhere.

This also meshes well with the fact that in postprocessing we also
mutate the hardlinked files (e.g. nuking `user.ostreemeta` xattrs) which
basically means the pkgcache repo is not actually usable as a cache.