libpriv/kernel: Pass through `DRACUT_NO_XATTR` when running dracut
Since f4aecb9b ("rust/bwrap: don't swallow STDERR when running
commands"), we now have the stderr from dracut, which reveals that it
spews lots of warnings about being unable to copy xattrs when running
from an `rpm-ostree compose rootfs`.

I think this comes from the `user.ostreemeta` xattr which hasn't been
stripped yet.

It'd be better anyway to do this stripping within the rpm-ostree core I
think, but for now it's actually trivial to work around this by setting
`DRACUT_NO_XATTR`.

I'm pretty sure we could set this unconditionally, but out of an
abundance of caution, we only do this for now from the
`rpm-ostree compose rootfs` path where we know we want this.

Fixes: https://gitlab.com/fedora/bootc/tracker/-/issues/66