core: Run sysusers after doing passwd/group layering dance
For now, we'll just treat sysusers entries from RPM packages like we do
scriptlets that `useradd`/`groupadd`; that is, we want them to happen
at compose time and go into altfiles in case those same sysusers own
content in the commit.

All we need to do to make that happen is to run `systemd-sysusers`
_after_ we do the `/etc/passwd` <--> `/usr/lib/passwd` switcheroo so
that the new entries go into what will become `/usr/lib/passwd`.

And all we need to do that is to just move down the sysusers execution
a bit.

Fixes: https://github.com/coreos/rpm-ostree/issues/5365