core: Neuter sysusers when writing rpmdb
I noticed while hacking on #5365 that during the rpmdb writing, librpm
was actually re-executing systemd-sysusers *from the host context* which
is not at all what we want.

Apparently, `RPMTRANS_FLAG_JUSTDB` doesn't imply this and we need to
explicitly also pass `RPMTRANS_FLAG_NOSYSUSERS`. That flag doesn't exist
in el9, so add a compile-time conditional for it.

This fixes the issue for new systems, but people who have upgraded
to f42 and overlaid packages with sysusers entries will have new entries
in `/etc/passwd` and `/etc/group` files because of this. And this can
cause problems now if the UIDs chosen were different because the `/etc`
entries will take precedence over nss-altfiles even though owned content
will match nss-altfiles.

In practice, I think since #5365 breaks exactly those use cases
where the sysusers entries own content, we don't have to worry about
that subcase. But for sysusers entries that _don't_ own content, the
transaction would go through and so there could still be UID conflicts
there.

I guess we'll need to figure out if to somehow try to fix this or just
issue a PSA about it.