passwd: Also save and restore shadow files
When doing the passwd/group layering dance, we need to take care of also
moving the canonical shadow files from the base layer out of the way the
same way we do passwd and group.

All those files are coherent with each other so it doesn't make sense to
move some and not others. Otherwise, when we move the `/usr/lib` version
of passwd and group into `/etc` as part of the dance and we try to run
systemd-sysusers, it can be confused by the fact that the shadow files
seemingly already have some entries while passwd and group do not.

In the end those `/etc` files don't really matter because they're almost
guaranteed to be dirtied on local systems. So the canonical copy in the
image will never actually be used.