Add support for skipping lua scripts via a magic comment
As the doc updates say, in many use cases for lua scripts
we have never needed them and just wanted to ignore them.

This gives people a way to denote that in the script itself,
without us having to change the client.

This doesn't fix https://github.com/coreos/rpm-ostree/issues/749
but should greatly lessen the pain from it.

Signed-off-by: Colin Walters <walters@verbum.org>