treefile: Add `sysusers:` knob
In the work to rebase FCOS/RHCOS to fedora-bootc/rhel-bootc, I want
to make it possible to opt out of the passwd/group bits baked in the
minimal manifests (the problem there of course goes back to UID/GID
drift issues -- this has caused issues as well on the IoT/Edge side).
The current passwd/group files in the minimal manifest are both outdated
and incomplete (most of it can in fact trace its lineage all the way
back to Fedora Atomic Host).

But more generally, in the FROM scratch flow, users _should_ have the
ability to customize which user gets which UID. And one could imagine
an option like `bootc-base-imagectl --passwd mypasswd --group mygroup`
but (1) while the format is well-known, using it as an interface
for building root filesystems is less so, and for example has no
correspondence in the dnf world, and (2) such an interface implies
taking ownership of all the entries.

A more universal config file for this of course is systemd-sysusers. And
recently, rpm-ostree learned in #5334 to run systemd-sysusers prior to
running any scriptlets for parity with similar functionality in RPM.

So it basically works today in Fedora to completely ditch the monolithic
passwd/group stuff and instead fixate UIDs via sysusers dropins. And
then, combining that with `bootc-base-imagectl --add-dir`, this becomes
possible to feed into the FROM scratch flow.

We want to enable this flow in RHEL as well. All we need really
is to force rpm-ostree to run systemd-sysusers even if the
`%__systemd_sysusers` macro is undefined.

Add a new experimental `sysusers:` knob to the treefile. Currently, it
takes one of two values:
- `compose-auto`: `systemd-sysusers` is run at compose time if the
  `%__systemd_sysusers` RPM macro is defined (status quo)
- `compose-forced`: `systemd-sysusers` is always run at compose time,
  even if the `%__systemd_sysusers` RPM macro is undefined

And the net result of this then is it moves the emphasis
further to systemd-sysusers as a way of ensuring UID coherence
across builds and e.g. all the flexibility it implies around
configurability/composition/overrides etc...

(Worth noting that the systemd %post script actually always runs
systemd-sysusers currently. But we can't rely on this because it's
itself part of the RPM transaction and so has no effect on packages
which have run before.)

Then, for the bootc-base-imagectl flow, we can add a new `--sysusers`
switch which sets this new option and crucially disables all the
`check-passwd`/`check-group` logic.

This is related to the nss-altfiles to sysusers migration though
also somewhat independent. With `compose-forced`, the split back into
`/usr/lib/{passwd,group}` still happens at the end of the compose.

However I still went with the `sysusers:` enum and left room for other
strategies to be implemented, since any strategy around nss-altfiles
migration must take into account what semantics are desired from the
systemd-sysusers run at compose time.