compose: Add `--label` for build-chunked-oci
This is trivial to do since the underlying `container-encapsulate` we do
already supports this.

What becomes really interesting about this though is that in a container
build flow where one is using `build-chunked-oci` with the
`FROM: ./out.ociarchive` trick, it then becomes possible to
_dynamically_ (i.e. from within the build process itself) choose whether
to add a label, or determine what value a label should have, which is a
common frustration in container image building.