treefile,compose: add advisories-metadata option
Similar to repo-metadata, advisories-metadata option allow to set
if you should store advisories information in the commit metadata,
in the detached metadata, or not at all.

This help with reproducibility has advisories might evolve separetely
from the packages, and depending on how you create your mirrors
advisories might be missing.